requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.4.0
torchvision==0.19.0
pillow==10.4.0
imageio==2.36.1
imageio-ffmpeg==0.5.1
tqdm==4.67.1
easydict==1.13
opencv-python-headless==4.10.0.84
scipy==1.14.1
rembg==2.0.60
onnxruntime==1.20.1
trimesh==4.5.3
xatlas==0.0.9
pyvista==0.44.2
pymeshfix==0.17.0
igraph==0.11.8
git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8
xformers==0.0.27.post2
spconv-cu120
transformers==4.46.3
gradio_litmodel3d==0.0.1
stpyvista
python-dotenv
streamlit
huggingface_hub
diffusers[torch]